Taking out clusteringpath and better handling of  intervals to restrict or exclude from the analysis
This updates the CellExplorer file structure to only work in the basepath folder. the clusteringpath are only relevant now for where to load the spike data from. This simplifies the file handling substantial and makes it more in line with buzcode.
Spikes can now re restricted to specific intervals and intervals can be excluded on a metric level, meaning that the user can specify which metrics to pull from all spikes and which to pull from the spikes excluding manipulation intervals.