Change a message into a warning to make it easier to find
When executing a function which calls "basenameFromBasepath" it was very hard to find where this message was coming from. It was also not obvious if the issue "Failed to find basepath files" prevented the code from executing or not. Making the message into a warning solves both problems (easy to find which function called it, because Matlab reports the line where the warning is located, and also makes it clear that the code ran anyway).