fix issue of cell_metric file being saved improperly
Issue: whenever cell_metrics is updated through the gui, e.g. when adjusting monosynaptic connections, the updated file was saved with an erroneous format.

Instead of containing the "cell_metrics" structure, the line was saving an embedded "cell_metrics.cell_metrics" structure. This change of format prevented further use of CellExplorer.

This fix is making sure that the "cell_metrics" structure that is saved is not embedded into a parent structure, and it is saved in the same format it was loaded in.